Refactor Blarney to identify Netlist manipulation module and Backend module
