+++ Haskell/Blarney/Backend/Verilog.hs
@@ -1,21 +1,21 @@
+{-# LANGUAGE NoRebindableSyntax #-}
+
 {-|
-Module      : Blarney.Core.Verilog
+Module      : Blarney.Backend.Verilog
 Description : Verilog generation
 Copyright   : (c) Matthew Naylor, 2019
-              (c) Alexandre Joannou, 2019
+              (c) Alexandre Joannou, 2019-2020
 License     : MIT
 Maintainer  : mattfn@gmail.com
 Stability   : experimental
 
-Convert Blarney functions to Verilog modules.
+Convert Blarney Netlist to Verilog modules.
 -}
 
-module Blarney.Core.Verilog
-  ( writeVerilogModule     -- Generate Verilog module
-  , writeVerilogTop        -- Generate Verilog top-level module
-  , writeVerilogModuleOpts -- Generate Verilog module (with options)
-  , writeVerilogTopOpts    -- Generate Verilog top-level module (with options)
-  ) where
+module Blarney.Backend.Verilog (
+  genVerilogModule -- Generate Verilog module
+, genVerilogTop    -- Generate Verilog top-level module
+) where
 
 -- Standard imports
 import Prelude hiding ((<>))
@@ -30,109 +30,69 @@ import Numeric (showHex)
 import Data.Array.IArray
 
 -- Blarney imports
-import Blarney.Core.Net
-import Blarney.Core.Opts
-import Blarney.Core.Prim
 import Blarney.Core.Utils
-import Blarney.Core.Module
-import Blarney.Core.Interface
-import Blarney.Core.IfThenElse
+import Blarney.Netlist
 
 -- Toplevel API
 --------------------------------------------------------------------------------
 
--- | Convert given Blarney function to a Verilog module
-writeVerilogModuleOpts :: Modular a
-                       => Opts       -- ^ Options
-                       -> a          -- ^ Blarney function
-                       -> String     -- ^ Module name
-                       -> String     -- ^ Output directory
-                       -> IO ()
-writeVerilogModuleOpts opts top mod dir =
-    do system ("mkdir -p " ++ dir)
-       nl <- netlist opts (makeModule top)
-       writeVerilog fileName mod nl
-  where
-    fileName = dir ++ "/" ++ mod ++ ".v"
+-- | Convert given Blarney Netlist to a Verilog module
+genVerilogModule :: Netlist -- ^ Blarney netlist
+                 -> String  -- ^ Module name
+                 -> String  -- ^ Output directory
+                 -> IO ()
+genVerilogModule nl mod dir =
+  do system ("mkdir -p " ++ dir)
+     writeVerilog fileName mod nl
+  where fileName = dir ++ "/" ++ mod ++ ".v"
 
--- | Convert given Blarney function to a Verilog module using default options
-writeVerilogModule :: Modular a
-                   => a          -- ^ Blarney function
-                   -> String     -- ^ Module name
-                   -> String     -- ^ Output directory
-                   -> IO ()
-writeVerilogModule top mod dir = do
-  (opts, unused) <- getOpts
-  writeVerilogModuleOpts opts top mod dir
-
--- | Convert given Blarney function to a top-level Verilog module
-writeVerilogTopOpts :: Opts       -- ^ Options
-                    -> Module ()  -- ^ Blarney module
-                    -> String     -- ^ Top-level module name
-                    -> String     -- ^ Output directory
-                    -> IO ()
-writeVerilogTopOpts opts top mod dir =
-    do nl <- netlist opts top
-       system ("mkdir -p " ++ dir)
-       writeVerilog (dir ++ "/" ++ mod ++ ".v") mod nl
-       writeFile (dir ++ "/" ++ mod ++ ".cpp") simCode
-       writeFile (dir ++ "/" ++ mod ++ ".mk") makefileIncCode
-       writeFile (dir ++ "/Makefile") makefileCode
-  where
-    fileName = dir ++ "/" ++ mod ++ ".v"
-
-    simCode =
-      unlines [
-        "// Generated by Blarney"
-      , "#include <verilated.h>"
-      , "#include \"V" ++ mod ++ ".h\""
-      , "V" ++ mod ++ " *top;"
-      , "vluint64_t main_time = 0;"
-      , "// Called by $time in Verilog"
-      , "double sc_time_stamp () {"
-      , "  return main_time;"
-      , "}"
-      , "int main(int argc, char** argv) {"
-      , "  Verilated::commandArgs(argc, argv);"
-      , "  top = new V" ++ mod ++ ";"
-      , "  while (!Verilated::gotFinish()) {"
-      , "    top->clock = 0; top->eval();"
-      , "    top->clock = 1; top->eval();"
-      , "    main_time++;"
-      , "  }"
-      , "  top->final(); delete top; return 0;"
-      , "}"
-      ]
-
-    makefileIncCode =
-      unlines [
-        "all: " ++ mod
-      , mod ++ ": *.v *.cpp"
-      , "\tverilator -cc " ++ mod ++ ".v " ++ "-exe "
-                           ++ mod ++ ".cpp " ++ "-o " ++ mod
-                           ++ " -Wno-UNSIGNED"
-                           ++ " -y $(BLARNEY_ROOT)/Verilog"
-                           ++ " --x-assign unique"
-                           ++ " --x-initial unique"
-      , "\tmake -C obj_dir -j -f V" ++ mod ++ ".mk " ++ mod
-      , "\tcp obj_dir/" ++ mod ++ " ."
-      , "\trm -rf obj_dir"
-      , ".PHONY: clean clean-" ++ mod
-      , "clean: clean-" ++ mod
-      , "clean-" ++ mod ++ ":"
-      , "\trm -f " ++ mod
-      ]
-
-    makefileCode = "include *.mk"
-
--- | Convert Blarney function to top-level Verilog module with default options
-writeVerilogTop :: Module ()  -- ^ Blarney module
-                -> String     -- ^ Top-level module name
-                -> String     -- ^ Output directory
-                -> IO ()
-writeVerilogTop top mod dir = do
-  (opts, unused) <- getOpts
-  writeVerilogTopOpts opts top mod dir
+-- | Convert given Blarney Netlist to a top-level Verilog module
+genVerilogTop :: Netlist -- ^ Blarney module
+              -> String  -- ^ Top-level module name
+              -> String  -- ^ Output directory
+              -> IO ()
+genVerilogTop nl mod dir =
+  do genVerilogModule nl mod dir
+     writeFile (dir ++ "/" ++ mod ++ ".cpp") simCode
+     writeFile (dir ++ "/" ++ mod ++ ".mk") makefileIncCode
+     writeFile (dir ++ "/Makefile") makefileCode
+  where simCode = unlines [
+            "// Generated by Blarney"
+          , "#include <verilated.h>"
+          , "#include \"V" ++ mod ++ ".h\""
+          , "V" ++ mod ++ " *top;"
+          , "vluint64_t main_time = 0;"
+          , "// Called by $time in Verilog"
+          , "double sc_time_stamp () {"
+          , "  return main_time;"
+          , "}"
+          , "int main(int argc, char** argv) {"
+          , "  Verilated::commandArgs(argc, argv);"
+          , "  top = new V" ++ mod ++ ";"
+          , "  while (!Verilated::gotFinish()) {"
+          , "    top->clock = 0; top->eval();"
+          , "    top->clock = 1; top->eval();"
+          , "    main_time++;"
+          , "  }"
+          , "  top->final(); delete top; return 0;"
+          , "}" ]
+        makefileIncCode = unlines [
+            "all: " ++ mod
+          , mod ++ ": *.v *.cpp"
+          , "\tverilator -cc " ++ mod ++ ".v " ++ "-exe "
+                               ++ mod ++ ".cpp " ++ "-o " ++ mod
+                               ++ " -Wno-UNSIGNED"
+                               ++ " -y $(BLARNEY_ROOT)/Verilog"
+                               ++ " --x-assign unique"
+                               ++ " --x-initial unique"
+          , "\tmake -C obj_dir -j -f V" ++ mod ++ ".mk " ++ mod
+          , "\tcp obj_dir/" ++ mod ++ " ."
+          , "\trm -rf obj_dir"
+          , ".PHONY: clean clean-" ++ mod
+          , "clean: clean-" ++ mod
+          , "clean-" ++ mod ++ ":"
+          , "\trm -f " ++ mod ]
+        makefileCode = "include *.mk"
 
 writeVerilog :: String -> String -> Netlist -> IO ()
 writeVerilog fileName modName netlist = do
