+++ Haskell/Blarney/Core/Net.hs
@@ -6,7 +6,7 @@
 Module      : Blarney.Core.Net
 Description : Net primitive for Netlist construction
 Copyright   : (c) Matthew Naylor, 2019
-              (c) Alexandre Joannou, 2019
+              (c) Alexandre Joannou, 2019-2020
 License     : MIT
 Maintainer  : mattfn@gmail.com
 Stability   : experimental
@@ -18,27 +18,18 @@ that can then be rendered as Verilog or in other formats...
 
 module Blarney.Core.Net (
   Net(..)        -- 'Net' type to represent 'Netlist' nodes
-, Netlist        -- 'Netlist' type to represent a circuit
+, WireId         -- 'WireId' type to uniquely identify wires
 , NetInput(..)   -- 'NetInput' type to represent inputs to 'Net's
 , MNetlist       -- 'MNetlist' type, mutable netlist
-, readNet        -- Helper function to read a 'Net' out of an 'MNetlist'
-, WireId         -- 'WireId' type to uniquely identify wires
-, netlistPasses  -- Toplevel function for 'Netlist' transformation passes
+, Netlist        -- 'Netlist' type to represent a circuit
 ) where
 
 import Prelude
 import Data.Maybe
-import Data.IORef
 import Data.Array
 import Data.Array.IO
-import Control.Monad
-import Data.Set (insert, toList)
-import Data.List (intercalate)
-import qualified Data.Bits as B
 
-import Blarney.Core.Opts
 import Blarney.Core.Prim
-import Blarney.Core.IfThenElse
 
 -- General type definitions and helpers
 --------------------------------------------------------------------------------
@@ -65,439 +56,8 @@ data NetInput = InputWire WireId
               | InputTree Prim [NetInput]
               deriving Show
 
--- | A 'Netlist', represented as an 'Array InstId (Maybe Net)'
-type Netlist = Array InstId (Maybe Net)
-
 -- | A helper type for mutable 'Netlist'
 type MNetlist = IOArray InstId (Maybe Net)
 
--- | A helper function to tell if a 'Net' is a netlist root
-netIsRoot :: Net -> Bool
-netIsRoot Net{netPrim=prim} = primIsRoot prim
-
--- | A helper function to tell if a 'Net' should not be optimised away
-netDontKill :: Net -> Bool
-netDontKill Net{netPrim=prim} = primDontKill prim
-
--- | A helper function to read a 'Net' from a 'MNetlist'
-readNet :: MNetlist -> InstId -> IO Net
-readNet nl instId = fromMaybe (error "encountered InstId with no matching Net")
-                              <$> (readArray nl instId)
-
--- | A helper type for 'Net' reference counting
-type NetCounts = IOUArray InstId Int
-
--- | A 'Net' reference counting helper function
-countNetRef :: MNetlist -> IO NetCounts
-countNetRef arr = do
-  bounds <- getBounds arr
-  refCounts <- newArray bounds 0
-  es <- getElems arr
-  -- count references for each Net
-  let innerCount (InputWire (instId, _)) = do
-        cnt <- readArray refCounts instId
-        writeArray refCounts instId (cnt + 1)
-      innerCount (InputTree _ inpts) = mapM_ innerCount inpts
-  forM_ [e | Just e <- es] $ \net -> mapM_ innerCount (netInputs net)
-  -- return reference counts
-  return refCounts
-
--- | Repeat computation until a predicate holds
-untilM :: Monad m => (a -> Bool) -> m a -> m a
-untilM pred act =
-  act >>= \x -> if pred x then return x else untilM pred act
-
--- Netlist transformation passes
---------------------------------------------------------------------------------
-
--- | propagate names through the Netlist
-propagateNames :: MNetlist -> IO ()
-propagateNames nl = do
-  bounds <- getBounds nl
-  visited :: IOUArray InstId Bool <- newArray bounds False
-  -- Push destination name down through netlist
-  let visit destName instId = do
-      isVisited <- readArray visited instId
-      if not isVisited then do
-        net@Net{ netPrim = prim
-               , netInputs = inpts
-               , netNameHints = hints
-               } <- readNet nl instId
-        let inpts' = [instId | x@(InputWire (instId, _)) <- inpts]
-        writeArray visited instId True
-        -- Detect new destination and update destination name in recursive call
-        if isDest prim then mapM_ (visit $ bestName net) inpts'
-        else do
-          let newHints = insert (NmSuffix 10 destName) hints
-          writeArray nl instId $ Just net{netNameHints = newHints}
-          mapM_ (visit destName) inpts'
-      else return ()
-  --
-  pairs <- getAssocs nl -- list of nets with their index
-  forM_ [i | x@(i, Just n) <- pairs, netIsRoot n] $ \instId -> do
-    net <- readNet nl instId
-    visit (bestName net) instId
-  --
-  where bestName Net{ netPrim = prim
-                    , netNameHints = hints
-                    , netInstId = instId
-                    } = "DEST_" ++ nm
-                        where nm = if null nms
-                                   then primStr prim ++ "_id" ++ show instId
-                                   else head nms
-                              nms = [y | x@(NmRoot _ y) <- toList hints]
-        --
-        isDest Register{}     = True
-        isDest RegisterEn{}   = True
-        isDest BRAM{}         = True
-        isDest TrueDualBRAM{} = True
-        isDest Custom{}       = True
-        isDest Input{}        = True
-        isDest Output{}       = True
-        isDest Display{}      = True
-        isDest Finish         = True
-        isDest TestPlusArgs{} = True
-        isDest RegFileMake{}  = True
-        isDest RegFileRead{}  = True
-        isDest RegFileWrite{} = True
-        isDest _ = False
-
--- pattern helper to identify constant InputTree NetInputs
-pattern Lit i <- InputTree (Const _ i) []
--- | Helper to evaluate constant Net
-evalConstNet :: Net -> (Net, Bool)
--- Add
-evalConstNet n@Net{ netPrim = Add w, netInputs = [Lit 0, x] } =
-  (n { netPrim = Identity w, netInputs = [x] }, True)
-evalConstNet n@Net{ netPrim = Add w, netInputs = [x, Lit 0] } =
-  (n { netPrim = Identity w, netInputs = [x] }, True)
-evalConstNet n@Net{ netPrim = Add w, netInputs = [Lit a0, Lit a1] } =
-  (n { netPrim = Const w (a0 + a1), netInputs = [] }, True)
--- Sub -- TODO 0 - x could use a new negate primitive ?
-evalConstNet n@Net{ netPrim = Sub w, netInputs = [x, Lit 0] } =
-  (n { netPrim = Identity w, netInputs = [x] }, True)
-evalConstNet n@Net{ netPrim = Sub w, netInputs = [Lit a0, Lit a1] } =
-  (n { netPrim = Const w (a0 - a1), netInputs = [] }, True)
--- Mul
-evalConstNet n@Net{ netPrim = Mul w, netInputs = [Lit 0, _] } =
-  (n { netPrim = Const w 0, netInputs = [] }, True)
-evalConstNet n@Net{ netPrim = Mul w, netInputs = [_, Lit 0] } =
-  (n { netPrim = Const w 0, netInputs = [] }, True)
-evalConstNet n@Net{ netPrim = Mul w, netInputs = [Lit 1, x] } =
-  (n { netPrim = Identity w, netInputs = [x] }, True)
-evalConstNet n@Net{ netPrim = Mul w, netInputs = [x, Lit 1] } =
-  (n { netPrim = Identity w, netInputs = [x] }, True)
-evalConstNet n@Net{ netPrim = Mul w, netInputs = [Lit a0, Lit a1] } =
-  (n { netPrim = Const w (a0 * a1), netInputs = [] }, True)
--- Div
-evalConstNet n@Net{ netPrim = Div w, netInputs = [Lit 0, _] } =
-  (n { netPrim = Const w 0, netInputs = [] }, True)
-evalConstNet n@Net{ netPrim = Div w, netInputs = [x, Lit 1] } =
-  (n { netPrim = Identity w, netInputs = [x] }, True)
-evalConstNet n@Net{ netPrim = Div w, netInputs = [Lit a0, Lit a1] } =
-  (n { netPrim = Const w (a0 `div` a1), netInputs = [] }, True)
--- Mod
-evalConstNet n@Net{ netPrim = Mod w, netInputs = [Lit 0, _] } =
-  (n { netPrim = Const w 0, netInputs = [] }, True)
-evalConstNet n@Net{ netPrim = Mod w, netInputs = [x, Lit 1] } =
-  (n { netPrim = Const w 0, netInputs = [] }, True)
-evalConstNet n@Net{ netPrim = Mod w, netInputs = [Lit a0, Lit a1] } =
-  (n { netPrim = Const w (a0 `mod` a1), netInputs = [] }, True)
--- Not
-evalConstNet n@Net{ netPrim = Not w, netInputs = [Lit a0] } =
-  (n { netPrim = Const w ((2^w-1) `B.xor` a0), netInputs = [] }, True)
--- And
-evalConstNet n@Net{ netPrim = And w, netInputs = [Lit 0, x] } =
-  (n { netPrim = Const w 0, netInputs = [] }, True)
-evalConstNet n@Net{ netPrim = And w, netInputs = [x, Lit 0] } =
-  (n { netPrim = Const w 0, netInputs = [] }, True)
-evalConstNet n@Net{ netPrim = And w, netInputs = [Lit v, x] } | v == 2^w-1 =
-  (n { netPrim = Identity w, netInputs = [x] }, True)
-evalConstNet n@Net{ netPrim = And w, netInputs = [x, Lit v] } | v == 2^w-1 =
-  (n { netPrim = Identity w, netInputs = [x] }, True)
-evalConstNet n@Net{ netPrim = And w, netInputs = [Lit a0, Lit a1] } =
-  (n { netPrim = Const w (a0 B..&. a1), netInputs = [] }, True)
--- Or
-evalConstNet n@Net{ netPrim = Or w, netInputs = [Lit 0, x] } =
-  (n { netPrim = Identity w, netInputs = [x] }, True)
-evalConstNet n@Net{ netPrim = Or w, netInputs = [x, Lit 0] } =
-  (n { netPrim = Identity w, netInputs = [x] }, True)
-evalConstNet n@Net{ netPrim = Or w, netInputs = [Lit v, x] } | v == 2^w-1 =
-  (n { netPrim = Const w (2^w-1), netInputs = [] }, True)
-evalConstNet n@Net{ netPrim = Or w, netInputs = [x, Lit v] } | v == 2^w-1 =
-  (n { netPrim = Const w (2^w-1), netInputs = [] }, True)
-evalConstNet n@Net{ netPrim = Or w, netInputs = [Lit a0, Lit a1] } =
-  (n { netPrim = Const w (a0 B..|. a1), netInputs = [] }, True)
--- Xor
-evalConstNet n@Net{ netPrim = Xor w, netInputs = [Lit 0, x] } =
-  (n { netPrim = Identity w, netInputs = [x] }, True)
-evalConstNet n@Net{ netPrim = Xor w, netInputs = [x, Lit 0] } =
-  (n { netPrim = Identity w, netInputs = [x] }, True)
-evalConstNet n@Net{ netPrim = Xor w, netInputs = [Lit v, x] } | v == 2^w-1 =
-  (n { netPrim = Not w, netInputs = [x] }, True)
-evalConstNet n@Net{ netPrim = Xor w, netInputs = [x, Lit v] } | v == 2^w-1 =
-  (n { netPrim = Not w, netInputs = [x] }, True)
-evalConstNet n@Net{ netPrim = Xor w, netInputs = [Lit a0, Lit a1] } =
-  (n { netPrim = Const w (a0 `B.xor` a1), netInputs = [] }, True)
--- ShiftLeft
-evalConstNet n@Net{ netPrim = ShiftLeft _ w, netInputs = [Lit 0, _] } =
-  (n { netPrim = Const w 0, netInputs = [] }, True)
-evalConstNet n@Net{ netPrim = ShiftLeft _ w, netInputs = [x, Lit 0] } =
-  (n { netPrim = Identity w, netInputs = [x] }, True)
-evalConstNet n@Net{ netPrim = ShiftLeft _ w, netInputs = [Lit a0, Lit a1] } =
-  (n { netPrim = Const w (a0 `B.shiftL` fromInteger (a1))
-     , netInputs = [] }, True)
--- TODO XXX FORCE UNSIGNED FOR NON ARITHMETIC SHIFTS
---ev Net{ netPrim = ShiftRight w, netInputs = [a0, a1] } =
---  n { netPrim   = Const w (inV a0 `B.shiftR` fromInteger (inV a1))
---    , netInputs = [] }
---ev Net{ netPrim = ArithShiftRight w, netInputs = [a0, a1] } =
---  n { netPrim   = Const w (inV a0 `B.shiftR` fromInteger (inV a1))
---    , netInputs = [] }
--- TODO XXX FORCE UNSIGNED FOR NON ARITHMETIC SHIFTS
--- Comparisons
-evalConstNet n@Net{ netPrim = Equal _, netInputs = [Lit a0, Lit a1] } =
-  (n { netPrim = Const 1 (if a0 == a1 then 1 else 0), netInputs = [] }, True)
-evalConstNet n@Net{ netPrim = NotEqual _, netInputs = [Lit a0, Lit a1] } =
-  (n { netPrim = Const 1 (if a0 /= a1 then 1 else 0), netInputs = [] }, True)
-evalConstNet n@Net{ netPrim = LessThan _, netInputs = [Lit a0, Lit a1] } =
-  (n { netPrim = Const 1 (if a0 < a1 then 1 else 0), netInputs = [] }, True)
-evalConstNet n@Net{ netPrim = LessThanEq _, netInputs = [Lit a0, Lit a1] } =
-  (n { netPrim = Const 1 (if a0 <= a1 then 1 else 0), netInputs = [] }, True)
-evalConstNet n@Net{ netPrim = ReplicateBit w, netInputs = [Lit a0] } =
-  (n { netPrim = Const w (if a0 == 1 then 2^w-1 else 0), netInputs = [] }, True)
--- TODO XXX
---ev Net{ netPrim = ZeroExtend _ w1, netInputs = [a0] } =
---  n { netPrim   = Const w1 (if inV a0 > 0 then inV a0 else --TODO)
---    , netInputs = [] }
---ev Net{ netPrim = SignExtend _ w1, netInputs = [a0] } =
---  n { netPrim   = Const w1 (if inV a0 > 0 then inV a0 else --TODO)
---    , netInputs = [] }
--- TODO XXX
--- SelectBits
-evalConstNet n@Net{ netPrim = SelectBits w hi lo, netInputs = [Lit a0] } =
-  (n { netPrim   = Const (hi-lo+1) ((a0 `B.shiftR` lo) B..&. (2^(hi-lo+1)-1))
-    , netInputs = [] }, True)
--- Concat
-evalConstNet n@Net{ netPrim = Concat w0 w1, netInputs = [Lit a0, Lit a1] } =
-  (n { netPrim = Const (w0+w1) ((a0 `B.shiftL` w1) B..|. a1)
-     , netInputs = [] }, True)
--- Mux
-evalConstNet n@Net{ netPrim = Mux _ w, netInputs = (Lit s):xs } =
-  (n { netPrim = Identity w, netInputs = [xs !! fromInteger s] }, True)
--- Identity
-evalConstNet n@Net{ netPrim = Identity w, netInputs = [Lit a0] } =
-  (n { netPrim   = Const w a0, netInputs = [] }, True)
--- Fall-through case, no change
-evalConstNet n = (n, False)
-
--- | Constant folding pass
-foldConstants :: MNetlist -> IO Bool
-foldConstants nl = do
-  pairs <- getAssocs nl -- list of nets with their index
-  changed <- newIORef False -- keep track of modifications to the 'Netlist'
-  -- Evaluate each constant 'Net' and update it in the 'Netlist'
-  forM_ [(a, b) | x@(a, Just b) <- pairs] $ \(idx, net) -> do
-    let (net', change) = evalConstNet net
-    when change $ do writeIORef changed True -- keep track of changes
-                     writeArray nl idx $ Just net' -- update 'Netlist'
-  -- finish pass
-  -- DEBUG HELP -- x <- readIORef changed
-  -- DEBUG HELP -- putStrLn $ "foldConstant pass changed? " ++ show x
-  readIORef changed -- return whether a change occured
-
--- | Constant propagation pass
-propagateConstants :: MNetlist -> IO Bool
-propagateConstants nl = do
-  pairs <- getAssocs nl -- list of nets with their index
-  changed <- newIORef False -- keep track of modifications to the 'Netlist'
-  -- Turn constant 'InputWire' for each 'Net' into constant 'InputTree'
-  forM_ [(a, b) | x@(a, Just b) <- pairs] $ \(idx, net) -> do
-    -- process each 'NetInput' for the current 'Net'
-    inputs' <- forM (netInputs net) $ \inpt -> do
-      case inpt of
-        InputWire (instId, _) -> do
-          inptNet <- readNet nl instId
-          -- keep track of change when transforming into an 'InputTree'
-          case netPrim inptNet of
-            p@(Const _ _)  -> writeIORef changed True >> return (InputTree p [])
-            p@(DontCare _) -> writeIORef changed True >> return (InputTree p [])
-            _              -> return inpt
-        _ -> return inpt
-    -- update the current 'Net' in the 'Netlist'
-    writeArray nl idx (Just net { netInputs = inputs' })
-  -- finish pass
-  -- DEBUG HELP -- x <- readIORef changed
-  -- DEBUG HELP -- putStrLn $ "propagateConstant pass changed? " ++ show x
-  readIORef changed
-
--- | Helper to inline a 'Net''s inputs
-inlineNetInput :: MNetlist -> NetCounts -> NetInput -> IO (NetInput, Bool)
-inlineNetInput nl nc inpt@(InputWire (instId, _)) = do
-  -- read ref count for our referenced 'Net'
-  cnt <- readArray nc instId
-  -- read netPrim and netInputs for our referenced 'Net'
-  Net{ netPrim = prim, netInputs = inpts } <- readNet nl instId
-  -- attempt inlining our referenced 'Net', and its inputs recursively, also
-  -- returning if inlining could happen as a Bool
-  if cnt == 1 && canInline prim then do
-    (inpts', changes) <- unzip <$> mapM
-      (\x -> if canInlineInput prim then do (x', _) <- inlineNetInput nl nc x
-                                            return (x', True)
-             else return (x, False)) inpts
-    return (InputTree prim inpts', or changes)
-  else return (inpt, False)
-inlineNetInput nl nc (InputTree prim inpts) = do
-  (inpts', changes) <- unzip <$> mapM (inlineNetInput nl nc) inpts
-  return $ (InputTree prim inpts', or changes)
-
--- | Single reference 'Net' inlining pass
-inlineSingleRefNet :: MNetlist -> IO Bool
-inlineSingleRefNet nl = do
-  refCounts <- countNetRef nl -- reference count for each 'Net'
-  pairs <- getAssocs nl -- list of nets with their index
-  changed <- newIORef False -- keep track of modifications to the 'Netlist'
-  -- Inline each "inlinable" (that is with a supported combinational underlying
-  -- primitive) 'Net'
-  forM_ [(a,b) | x@(a, Just b) <- pairs, canInlineInput (netPrim b)] $
-    \(idx, net) -> do
-      (netInputs', changes) <- unzip <$> mapM (inlineNetInput nl refCounts)
-                                              (netInputs net)
-      when (or changes) $ do -- on change, update 'Netlist' and keep track
-        writeArray nl idx (Just net { netInputs = netInputs' })
-        writeIORef changed True
-  -- finish pass
-  -- DEBUG HELP -- x <- readIORef changed
-  -- DEBUG HELP -- putStrLn $ "inlineNetInput pass changed? " ++ show x
-  readIORef changed
-
--- | Transform a 'Net' with non-zero-width output and at least one zero-width
---   input into an equivalent 'Net' with no reference to any other zero-width
---   output 'Net' as its input
-zeroWidthNetTransform :: Net -> (Net, Bool)
--- straight transformation cases
-zeroWidthNetTransform net@Net{ netPrim = Equal 0 } =
-  (net { netPrim = Const 1 1, netInputs = [] }, True)
-zeroWidthNetTransform net@Net{ netPrim = NotEqual 0 } =
-  (net { netPrim = Const 1 0, netInputs = [] }, True)
-zeroWidthNetTransform net@Net{ netPrim = LessThan 0 } =
-  (net { netPrim = Const 1 0, netInputs = [] }, True)
-zeroWidthNetTransform net@Net{ netPrim = LessThanEq 0 } =
-  (net { netPrim = Const 1 1, netInputs = [] }, True)
-zeroWidthNetTransform net@Net{ netPrim = ZeroExtend 0 w } =
-  (net { netPrim = Const w 0, netInputs = [] }, True)
-zeroWidthNetTransform net@Net{ netPrim = SignExtend 0 w } =
-  (net { netPrim = Const w 0, netInputs = [] }, True)
-zeroWidthNetTransform net@Net{ netPrim = SelectBits 0 hi lo } =
-  (net { netPrim = Const (hi-lo) 0, netInputs = [] }, True)
-zeroWidthNetTransform net@Net{ netPrim = Concat w0 w1, netInputs = [i0, i1] }
-  | w0 == 0 && w1 /= 0 = (net { netPrim = Identity w1, netInputs = [i1] }, True)
-  | w0 /= 0 && w1 == 0 = (net { netPrim = Identity w0, netInputs = [i0] }, True)
-  | otherwise = (net, False)
-zeroWidthNetTransform net@Net{ netPrim = Display args, netInputs = inpts } =
-  let bitArgs = [ ba | ba@(DisplayArgBit w) <- args ]
-      f (DisplayArgBit 0) i = (InputTree (Const 0 0) [], True)
-      f _ i = (i, False)
-      (tmps, changes) = unzip $ zipWith f bitArgs (tail inpts)
-      inpts' = head inpts : tmps
-  in (net { netInputs = inpts' }, or changes)
-zeroWidthNetTransform net@Net{ netPrim = prim@Custom{ customInputs = primIns
-                                                    , customOutputs = primOuts }
-                             , netInputs = netIns }
-  | any (\(_, x) -> x == 0) primIns ||
-    any (\(_, y) -> y == 0) primOuts =
-    ( net { netPrim = prim { customInputs = primIns'
-                           , customOutputs = primOuts' }
-          , netInputs = netIns' }
-    , True )
-  | otherwise = (net, False)
-  where ins = zip netIns primIns
-        ins' = [x | x@(_, (_, w)) <- ins, w /= 0]
-        (netIns', primIns') = unzip ins'
-        primOuts' = [x | x@(_, w) <- primOuts, w /= 0]
--- TODO currently unsupported cases that could be transformed
-zeroWidthNetTransform net@Net{ netPrim = BRAM { ramAddrWidth = 0 } } =
-  error "zeroWidthNetTransform unsupported on BRAM Prim"
-zeroWidthNetTransform net@Net{ netPrim = TrueDualBRAM { ramAddrWidth = 0 } } =
-  error "zeroWidthNetTransform unsupported on TrueDualBRAM Prim"
-zeroWidthNetTransform
-  net@Net{ netPrim = RegFileRead RegFileInfo{ regFileAddrWidth = 0 } } =
-    error "zeroWidthNetTransform unsupported on RegFileRead Prim"
--- do nothing cases for all others
-zeroWidthNetTransform net = (net, False)
-
--- | Tell if a root 'Net' can be optimized away during zero-witdh elimination,
--- that is a root 'Net' with no non-zero-width inputs or outputs
-zeroWidthRootNetEliminationRule Net{ netPrim = prim } =
-  primIsRoot prim && not (primDontKill prim) &&
-  not (any (\(_, w) -> w /= 0) ins || any (\(_, w) -> w /= 0) outs)
-  where ins = primInputs prim
-        outs = primOutputs prim
-
--- | Ignore 0-width Nets pass
-ignoreZeroWidthNet :: MNetlist -> IO Bool
-ignoreZeroWidthNet nl = do
-  pairs <- getAssocs nl -- list of nets with their index
-  changed <- newIORef False -- keep track of modifications to the 'Netlist'
-  -- For each 'Net' (in particular, those with a non-zero-width output) with at
-  -- least one zero-width input, transform it into a 'Net' with no reference to
-  -- any zero-width output 'Net'.
-  forM_ [(a,b) | x@(a, Just b) <- pairs] $ \(idx, net@Net{netPrim = prim}) -> do
-    let (net', netChanged) = zeroWidthNetTransform net
-    when netChanged $ do writeArray nl idx (Just net')
-                         writeIORef changed True
-  -- Remove each zero-width root 'Net'
-  forM_ [ i | x@(i, Just n) <- pairs
-            , zeroWidthRootNetEliminationRule n ] $ \idx -> do
-    writeArray nl idx Nothing
-    writeIORef changed True
-  -- finish pass
-  -- DEBUG HELP -- x <- readIORef changed
-  -- DEBUG HELP -- putStrLn $ "ignoreZeroWidthNet pass changed? " ++ show x
-  readIORef changed
-
--- | Dead Net elimination pass
-eliminateDeadNet :: MNetlist -> IO Bool
-eliminateDeadNet nl = do
-  refCounts <- countNetRef nl -- reference count for each 'Net'
-  pairs <- getAssocs nl -- list of nets with their index
-  changed <- newIORef False -- keep track of modifications to the 'Netlist'
-  -- kill Nets with a null reference count
-  forM_ [(a,b) | x@(a, Just b) <- pairs] $ \(idx, net) -> do
-    refCnt <- readArray refCounts idx
-    when (refCnt == 0 && (not . netIsRoot) net && (not . netDontKill) net) $ do
-      writeArray nl idx Nothing
-      writeIORef changed True
-  -- finish pass
-  -- DEBUG HELP -- x <- readIORef changed
-  -- DEBUG HELP -- putStrLn $ "eliminateDeadNet pass changed? " ++ show x
-  readIORef changed
-  where alsoDontKill Net{netPrim=Output _ _} = True
-        alsoDontKill Net{netPrim=RegFileWrite _} = True
-        alsoDontKill _ = False
-
--- | Run netlist transformation passes
-netlistPasses :: Opts -> MNetlist -> IO Netlist
-netlistPasses opts nl = do
-  -- remove 'Bit 0' instances
-  ignoreZeroWidthNet nl
-  -- netlist optimisation passes
-  when (optEnableSimplifier opts) $ do
-    let constElim i = do a <- foldConstants nl
-                         b <- propagateConstants nl
-                         -- DEBUG HELP
-                         -- putStrLn $ "constElim " ++ show (a || b)
-                         return $ a || b
-    -- DEBUG HELP -- putStrLn $ "about to untilM constElim"
-    untilM not $ constElim nl
-    -- DEBUG HELP -- putStrLn $ "about to inlineSingleRefNet"
-    inlineSingleRefNet nl
-    return ()
-  -- propagates existing names through the netlist
-  when (optEnableNamePropagation opts) $ propagateNames nl
-  -- eliminate 'Net' entries in the netlist for 'Net's that are no longer
-  -- referenced
-  -- DEBUG HELP -- putStrLn $ "about to eliminateDeadNet"
-  untilM not $ eliminateDeadNet nl
-  -- turn the final netlist immutable
-  freeze nl
+-- | A 'Netlist', represented as an 'Array InstId (Maybe Net)'
+type Netlist = Array InstId (Maybe Net)
